src/requirements.txt
# Django and REST Framework
Django>=5.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=4.3.0
drf-spectacular>=0.27.0

# Database
psycopg2-binary>=2.9.9
pgvector>=0.2.4

# Celery
celery>=5.3.6
redis>=5.0.1

# AI / ML
sentence-transformers>=2.2.2
transformers>=4.36.0
torch>=2.1.0
huggingface_hub>=0.20.0
safetensors>=0.4.0
tokenizers>=0.15.0
scikit-learn>=1.3.0

# PDF Processing
PyMuPDF>=1.23.0
reportlab>=4.0.0

# Notifications
twilio>=8.10.0

# Production server
gunicorn>=21.2.0

# Other dependencies
numpy>=1.26.0
scipy>=1.11.0
requests>=2.31.0
PyYAML>=6.0.1
python-dateutil>=2.8.2
certifi>=2023.11.17
urllib3>=2.1.0
